Refactor catalog.rs & add finish_streaming_job
